+++ MapboxSearch/sample/src/androidTest/java/com/mapbox/search/sample/ExampleActivitiesTest.kt
@@ -9,7 +9,7 @@ import com.mapbox.search.sample.api.CategorySearchJavaExampleActivity
 import com.mapbox.search.sample.api.CategorySearchKotlinExampleActivity
 import com.mapbox.search.sample.api.CustomIndexableDataProviderJavaExample
 import com.mapbox.search.sample.api.CustomIndexableDataProviderKotlinExample
-import com.mapbox.search.sample.api.DiscoverApiKotlinExampleActivity
+import com.mapbox.search.sample.api.DiscoverKotlinExampleActivity
 import com.mapbox.search.sample.api.FavoritesDataProviderJavaExample
 import com.mapbox.search.sample.api.FavoritesDataProviderKotlinExample
 import com.mapbox.search.sample.api.ForwardGeocodingBatchResolvingJavaExampleActivity
@@ -59,8 +59,8 @@ class ExampleActivitiesTest(private val clazz: Class<out Activity>) {
             AddressAutofillKotlinExampleActivity::class.java,
             AddressAutofillUiActivity::class.java,
 
-            DiscoverApiKotlinExampleActivity::class.java,
-            DiscoverApiUiActivity::class.java,
+            DiscoverKotlinExampleActivity::class.java,
+            DiscoverActivity::class.java,
 
             CustomThemeActivity::class.java,